feat(neovim): rename `tsserver` -> `ts_ls`
